+++ forge/src/main/java/com/slymask3/instantblocks/init/ForgeTiles.java
@@ -12,6 +12,6 @@ public class ForgeTiles implements ITileHelper {
         ModTiles.HARVEST = BlockEntityType.Builder.of(HarvestBlockEntity::new, ModBlocks.INSTANT_HARVEST).build(null);
         ModTiles.SCHEMATIC = BlockEntityType.Builder.of(SchematicBlockEntity::new, ModBlocks.INSTANT_SCHEMATIC).build(null);
         ModTiles.TREE = BlockEntityType.Builder.of(TreeBlockEntity::new, ModBlocks.INSTANT_TREE).build(null);
-        ModTiles.COLOR = BlockEntityType.Builder.of(ColorBlockEntity::new, ModBlocks.COLOR, ModBlocks.SKYDIVE_TP).build(null);
+        ModTiles.COLOR = BlockEntityType.Builder.of(ColorBlockEntity::new, ModBlocks.COLOR, ModBlocks.COLOR_LAYER, ModBlocks.SKYDIVE_TP).build(null);
     }
 }